requirements.txt
Flask==2.2.5
cryptography==41.0.2
python-dotenv==1.0.0
